refactor: Modal.js와 restaurantForm.js의 중복되는 부분을 합쳐 RestaurantForm.Modal.js 생성
